refactor: 강점 리포트, 커리어 로드맵 페이지 SSR로 전환
